+++ mne/inverse_sparse/mxne_inverse.py
@@ -556,7 +556,7 @@ def tf_mixed_norm(evoked, forward, noise_cov, alpha_space, alpha_time,
        "Time-Frequency Mixed-Norm Estimates: Sparse M/EEG imaging with
        non-stationary source activations",
        Neuroimage, Volume 70, pp. 410-422, 15 April 2013.
-       DOI: 10.1016/j.neuroimage.2012.12.051.
+       DOI: 10.1016/j.neuroimage.2012.12.051
 
     .. [2] A. Gramfort, D. Strohmeier, J. Haueisen, M. Hamalainen, M. Kowalski
        "Functional Brain Imaging with M/EEG Using Structured Sparsity in